python/audio/requirements.txt
openai-whisper>=20250625
torch>=2.0.0
torchaudio>=2.0.0
